move lsat route validation into its own middleware
